requirements.txt
fastapi
uvicorn
python-multipart
jinja2
google-search-results
python-dotenv